(minor) rename abbrevitateNumber to abbreviateNumbers
