+++ src/helpers/abbreviateNumbers.js
